// src/app/models/crud/cliente.ts
export class Cliente {
  id?: number = 0;
  tipoDocumento: string | null | undefined = "";
  documento: string | null | undefined = "";
  nombre: string | null | undefined = "";
  empresa: string | null | undefined = "";
  direccion: string | null | undefined = "";
  telefono: string | null | undefined = "";
  correo: string | null | undefined = "";
  departamento: string | null | undefined = "";
  municipio: string | null | undefined = "";
  estado?: string | null | undefined = "";
}